Use float variant of std::hypot

+++ gemrb/core/PathFinder.cpp
@@ -312,7 +312,7 @@ Path Map::FindPath(const Point& s, const Point& d, const unsigned int size, unsi
 		const int dxCross = smptDest.x - smptSource.x;
 		const int dyCross = smptDest.y - smptSource.y;
 		const int crossProduct = std::abs(xDist * dyCross - yDist * dxCross) >> 3;
-		const float distance = std::hypot(xDist, yDist);
+		const float distance = std::hypotf(xDist, yDist);
 		const float heuristic = HEURISTIC_WEIGHT * (distance + crossProduct);
 		const float estDist = distFromStart[smptChildIdx] + heuristic;
 		return estDist;
